purge request

+++ pin/models.py
@@ -519,7 +519,8 @@ class Post(models.Model):
         #                actor=self.user.id)
 
         super(Post, self).delete(*args, **kwargs)
-        PostCacheLayer(post_id=self.id).delete()
+        if settings.TUNING_CACHE:
+            PostCacheLayer(post_id=self.id).delete()
 
     def date_lt(self, date, how_many_days=15):
         lt_date = datetime.now() - timedelta(days=how_many_days)
@@ -664,7 +665,8 @@ class Post(models.Model):
         self.text = normalize_tags(self.text)
         # print "all save"
         super(Post, self).save(*args, **kwargs)
-        PostCacheLayer(post_id=self.id).post_change(self)
+        if settings.TUNING_CACHE:
+            PostCacheLayer(post_id=self.id).post_change(self)
         # print "after save - thumbnail "
 
     @models.permalink
@@ -1216,7 +1218,8 @@ class Comments(models.Model):
         comment_cache_name = "com_%d" % int(self.object_pk_id)
         cache.delete(comment_cache_name)
         super(Comments, self).save(*args, **kwargs)
-        PostCacheLayer(post_id=self.object_pk.id).comment_change(self.object_pk.cnt_comment)
+        if settings.TUNING_CACHE:
+            PostCacheLayer(post_id=self.object_pk.id).comment_change(self.object_pk.cnt_comment)
 
     @classmethod
     def add_comment(cls, sender, instance, created, *args, **kwargs):
@@ -1282,7 +1285,8 @@ class Comments(models.Model):
         comment_cache_name = "com_%d" % self.object_pk.id
         cache.delete(comment_cache_name)
         super(Comments, self).delete(*args, **kwargs)
-        PostCacheLayer(post_id=self.object_pk.id).delete_comment(self.object_pk.cnt_comment)
+        if settings.TUNING_CACHE:
+            PostCacheLayer(post_id=self.object_pk.id).delete_comment(self.object_pk.cnt_comment)
 
     @models.permalink
     def get_absolute_url(self):
